[webapp] Extract non-component exports
