+++ services/webapp/ui/src/contexts/TelegramContext.tsx
@@ -1,8 +1,6 @@
-import { createContext, useContext, ReactNode } from "react";
+import { ReactNode } from "react";
 import { useTelegram } from "@/hooks/useTelegram";
-
-// TelegramContext will hold values returned by useTelegram hook
-const TelegramContext = createContext<ReturnType<typeof useTelegram> | null>(null);
+import { TelegramContext } from "./telegram-context";
 
 export const TelegramProvider = ({ children }: { children: ReactNode }) => {
   const telegram = useTelegram();
@@ -11,13 +9,5 @@ export const TelegramProvider = ({ children }: { children: ReactNode }) => {
       {children}
     </TelegramContext.Provider>
   );
-};
-
-export const useTelegramContext = () => {
-  const context = useContext(TelegramContext);
-  if (!context) {
-    throw new Error("useTelegramContext must be used within TelegramProvider");
-  }
-  return context;
 };
