+++ services/webapp/ui/src/reminders/CreateReminder.tsx
@@ -4,7 +4,7 @@ import { MedicalButton, Sheet } from "@/components";
 import { cn } from "@/lib/utils";
 import { createReminder, updateReminder, getReminder } from "@/api/reminders";
 import { Reminder as ApiReminder } from "@sdk";
-import { useTelegramContext } from "@/contexts/TelegramContext";
+import { useTelegramContext } from "@/contexts/telegram-context";
 import { useToast } from "@/hooks/use-toast";
 import {
   normalizeReminderType,
